backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
web3==6.11.1
qrcode[pil]==7.4.2
Pillow==10.1.0
requests==2.31.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
